FIRM28: Rename files to match firmware guidelines
