+++ Manual_Tests/common/stm32/mcu_errors/MCUErrorsTest.c
