+++ Manual_Tests/common/stm32/mcu_info/MCUInfoTest.c
