+++ Manual_Tests/common/stm32/uart_rx/UARTRXTest.c
@@ -1,4 +1,4 @@
-#include <common/stm32/uart/log.h>
+#include <common/stm32/uart/Log.h>
 
 int main() {
     // Try to automatically detect board based on MCU UID
