+++ Manual_Tests/common/stm32/uart_tx/UARTTXTest.c
