+++ Manual_Tests/common/stm32/util/UtilTest.c
