+++ Manual_Tests/nucleo_g474re/g474re/G474RETest.c
