+++ Src/common/stm32/gpio/GPIOInput.h
@@ -5,10 +5,10 @@
  *      Author: bruno
  */
 
-#ifndef COMMON_STM32_GPIO_GPIO_INPUT_H_
-#define COMMON_STM32_GPIO_GPIO_INPUT_H_
+#ifndef COMMON_STM32_GPIO_GPIOINPUT_H_
+#define COMMON_STM32_GPIO_GPIOINPUT_H_
 
-#include <common/stm32/mcu/mcu.h>
+#include <common/stm32/mcu/MCU.h>
 #include <stdbool.h>
 
 typedef struct {
@@ -28,4 +28,4 @@ bool gpio_wait_for_state(GPIOInput* gpio, GPIO_PinState state,
 bool gpio_wait_for_low(GPIOInput* gpio, uint32_t timeout_ms);
 bool gpio_wait_for_high(GPIOInput* gpio, uint32_t timeout_ms);
 
-#endif /* COMMON_STM32_GPIO_GPIO_INPUT_H_ */
+#endif /* COMMON_STM32_GPIO_GPIOINPUT_H_ */
