+++ Src/common/stm32/mcu/HAL.h
