+++ Src/common/stm32/mcu/MCU.h
@@ -8,7 +8,7 @@
 #ifndef COMMON_STM32_MCU_MCU_H_
 #define COMMON_STM32_MCU_MCU_H_
 
-#include <common/stm32/mcu/hal.h>
+#include <common/stm32/mcu/HAL.h>
 
 // DEV_ID values
 typedef enum {
