+++ Src/common/stm32/uart/Log.h
@@ -8,7 +8,7 @@
 #ifndef COMMON_STM32_UART_LOG_H_
 #define COMMON_STM32_UART_LOG_H_
 
-#include <common/stm32/uart/uart.h>
+#include <common/stm32/uart/UART.h>
 #include <stdarg.h>
