+++ Src/common/stm32/util/Random.c
@@ -20,9 +20,9 @@
  * programs, not for anything cryptographically secure.
  */
 
-#include <common/stm32/util/random.h>
-#include <common/stm32/mcu/errors.h>
-#include <common/stm32/util/util.h>
+#include <common/stm32/mcu/Errors.h>
+#include <common/stm32/util/Random.h>
+#include <common/stm32/util/Util.h>
 #include <stdlib.h>
