+++ Src/common/stm32/util/Util.c
@@ -5,7 +5,7 @@
  *      Author: bruno
  */
 
-#include <common/stm32/util/util.h>
+#include <common/stm32/util/Util.h>
 
 
 /*
