+++ Src/common/stm32/util/Util.h
