+++ Src/nucleo_g474re/G474RE.h
@@ -1,14 +1,14 @@
 /*
- * h743zi2.h
+ * g474re.h
  *
  *  Created on: Jul. 25, 2021
  *      Author: bruno
  */
 
-#ifndef NUCLEO_H743ZI2_H743ZI2_H_
-#define NUCLEO_H743ZI2_H743ZI2_H_
+#ifndef NUCLEO_G474RE_G474RE_H_
+#define NUCLEO_G474RE_G474RE_H_
 
-#include <common/stm32/uart/log.h>
+#include <common/stm32/uart/Log.h>
 
 // Struct to represent the MCU on the dev kit and all peripherals/configuration
 typedef struct {
@@ -16,9 +16,9 @@ typedef struct {
     UART uart;
     Log log;
     GPIOInput blue_button;
-    GPIOOutput red_led;
-} H743ZI2;
+    GPIOOutput green_led;
+} G474RE;
 
-void h743zi2_init(H743ZI2* dev_kit);
+void g474re_init(G474RE* dev_kit);
 
-#endif /* NUCLEO_H743ZI2_H743ZI2_H_ */
+#endif /* NUCLEO_G474RE_G474RE_H_ */
