+++ Src/nucleo_h743zi2/H743ZI2.c
@@ -5,8 +5,8 @@
  *      Author: bruno
  */
 
-#include <nucleo_h743zi2/h743zi2.h>
-#include <nucleo_h743zi2/h743zi2_config.h>
+#include <nucleo_h743zi2/H743ZI2.h>
+#include <nucleo_h743zi2/H743ZI2Config.h>
 
 void h743zi2_init(H743ZI2* dev_kit) {
     mcu_init(&dev_kit->mcu, H743ZI2_BOARD);
